# assisted_team/swagger.yaml
openapi: 3.0.0
info:
  title: Aviasales Assisted Test Task
  version: 0.0.1
servers:
  - url: http://127.0.0.1:8080/api
    description: Main API endpoint

paths:
  /flights:
    get:
      summary: Returns a list of flights
      parameters:
        - name: source
          in: query
          description: (not used)3-letters IATA airport code of the departure port
          schema:
            type: string
            default: DXB
        - name: destination
          in: query
          description: (not used)3-letters IATA airport code of the arrival port
          schema:
            type: string
            default: BKK
      responses:
        '200':
          description: List of available travel Itineraries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Travel'
  /top:
    get:
      summary: Returns a list of top flights according to the given criteria
      parameters:
        - name: criteria
          in: query
          required: true
          description: "type of the criteria. One of: price, duration, optimal"
          schema:
            type: string
            default: optimal
        - name: sort
          in: query
          description: "sorting order. One of: desc, asc"
          schema:
            type: string
            default: desc
        - name: source
          in: query
          description: (not used)3-letters IATA airport code of the departure port
          schema:
            type: string
            default: DXB
        - name: destination
          in: query
          description: (not used)3-letters IATA airport code of the arrival port
          schema:
            type: string
            default: BKK
      responses:
        '200':
          description: List of available travel Itineraries
          content:
            application/json:
              schema:
                type: object
                properties:
                  travels:
                    type: array
                    items:
                      $ref: '#/components/schemas/Travel'
                  value:
                    type: string


components:
  schemas:
    Flight:
      type: object
      properties:
        source:
          type: string
        destination:
          type: string
        departure_time:
          type: string
          format: datetime
        arrival_time:
          type: string
          format: datetime
        carrier:
          type: string
        flight_number:
          type: string
    PriceInfo:
      type: array
      items:
        type: object
        properties:
          type:
            type: string
          price:
            type: number
    Travel:
      type: object
      properties:
        flights:
          type: array
          items:
            $ref: '#/components/schemas/Flight'
        price_currency:
          type: string
        price_information:
          type: array
          items:
            $ref: '#/components/schemas/PriceInfo'